[component] overhaul
